mesonpy: refactor module into a package
