+++ meson.build
@@ -7,4 +7,10 @@ if py.language_version().version_compare('< 3.6')
   error('Invalid Python version, only >= 3.6 is supported.')
 endif
 
-py.install_sources('mesonpy.py')
+py.install_sources(
+  'mesonpy/__init__.py',
+  'mesonpy/_compat.py',
+  'mesonpy/_tags.py',
+  'mesonpy/_util.py',
+  subdir: 'mesonpy',
+)
